extract campaign read model routes
